+++ src/Juvix/Core/Language.hs
@@ -5,7 +5,6 @@
 
 module Juvix.Core.Language
   ( module Juvix.Core.Language,
-    module Juvix.Core.Language.Builtins,
     module Juvix.Core.Language.Base,
   )
 where
@@ -15,22 +14,9 @@ where
 -}
 
 import Juvix.Core.Language.Base
-import Juvix.Core.Language.Builtins
 
 {---------------------------------------------------------------------------------}
 {- Program tree datatype -}
-
--- Consecutive symbol IDs for reachable user functions.
-type Symbol = Word
-
--- Tag of a constructor, uniquely identifying it. Tag values are consecutive and
--- separate from symbol IDs. We might need fixed special tags in Core for common
--- "builtin" constructors, e.g., unit, nat, lists, pairs, so that the code
--- generator can treat them specially.
-data Tag = BuiltinTag BuiltinDataTag | UserTag Word
-  deriving stock (Eq, Generic)
-
-instance Hashable Tag
 
 -- de Bruijn index
 type Index = Int
